# apps/console-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sharpn Console API",
    "version": "1.0.0",
    "description": "What the Console may ask of the Console API. Generated from the Zod schemas in `apps/console-api/src/contract`; edit those, not this."
  },
  "paths": {
    "/api/v1/session": {
      "post": {
        "operationId": "createSession",
        "summary": "Sign in",
        "description": "The one password, checked against the argon2id hash this service was deployed with. A success carries no body — the session is the cookie this sets, `HttpOnly`, `Secure`, `SameSite=Strict`, and there is no account to describe.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The password matched. The cookie is the session.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST` — The request body is not a Command the Console API can send: a key it does not know, a number out of range, a `null`, or nothing to change.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "`NOT_AUTHENTICATED` — No session, or one that has expired. Sign in again.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "`CSRF_REJECTED` — A mutation arrived without the CSRF token its session expects.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "`TOO_MANY_REQUESTS` — This browser is going too fast. Reads are not metered.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              },
              "Retry-After": {
                "description": "How long to wait before asking again, in seconds. Sent only with the codes on this status that can be waited out; several codes share a status, and most of them cannot.",
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "endSession",
        "summary": "Sign out",
        "description": "Clears the cookie. Answered the same way whether or not one was there to clear: there is no session store to consult, so there is nothing to disagree about.",
        "responses": {
          "204": {
            "description": "The cookie is cleared, or was never set.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            }
          },
          "403": {
            "description": "`CSRF_REJECTED` — A mutation arrived without the CSRF token its session expects.",
            "headers": {
              "x-correlation-id": {
                "$ref": "#/components/headers/x-correlation-id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginCommand": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "password"
        ],
        "additionalProperties": false,
        "description": "The one password, and nothing else: there is no account to name alongside it."
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BRIDGE_REJECTED_COMMAND",
                  "LIGHT_NOT_FOUND",
                  "GATEWAY_NOT_PAIRED",
                  "BRIDGE_UNREACHABLE",
                  "BRIDGE_TIMEOUT",
                  "MUTATION_OUTCOME_UNKNOWN",
                  "BRIDGE_BUSY",
                  "BRIDGE_UNSUPPORTED",
                  "GATEWAY_MISCONFIGURED",
                  "GATEWAY_ERROR",
                  "GATEWAY_UNREACHABLE",
                  "NOT_AUTHENTICATED",
                  "CSRF_REJECTED",
                  "TOO_MANY_REQUESTS",
                  "INVALID_REQUEST",
                  "COLOR_AND_TEMPERATURE_BOTH_SET"
                ]
              },
              "message": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              },
              "correlationId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "correlationId"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Every error, in one shape. `message` is the Console API's own words. `detail` is whoever refused, verbatim — the Gateway's prose, the Bridge's own description, or its own account of what was wrong with the body."
      }
    },
    "headers": {
      "x-correlation-id": {
        "description": "The Correlation ID this request was logged under. Minted here and never taken from the browser; the same value appears in every error envelope.",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    }
  }
}